feat: Convert withMDXComponents to use hooks (#417)
