+++ examples/create-react-app/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "dependencies": {
     "mdx.macro": "^0.2.7",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
     "react-scripts": "2.1.1"
   },
   "scripts": {
