+++ examples/parcel/package.json
@@ -7,8 +7,8 @@
   },
   "dependencies": {
     "@mdx-js/tag": "^0.16.1",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3"
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0"
   },
   "devDependencies": {
     "@babel/core": "^7.1.6",
