+++ examples/razzle/package.json
@@ -13,7 +13,7 @@
     "express": "^4.16.4",
     "razzle": "^2.4.0",
     "razzle-plugin-mdx": "^2.4.0",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3"
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0"
   }
 }
