+++ packages/loader/package.json
@@ -38,7 +38,7 @@
     "babel-preset-stage-0": "^6.24.1",
     "jest": "^23.0.0",
     "memory-fs": "^0.4.1",
-    "react": "^16.3.2",
+    "react": "^16.8.0",
     "webpack": "^4.5.0"
   },
   "scripts": {
