+++ packages/mdx/package.json
@@ -51,8 +51,8 @@
     "hast-util-select": "^3.0.0",
     "jest": "^23.6.0",
     "prettier": "^1.14.2",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
     "rehype-katex": "^1.1.1",
     "remark-math": "^1.0.4"
   },
