+++ packages/parcel-plugin-mdx/package.json
@@ -32,12 +32,12 @@
   },
   "peerDependencies": {
     "@mdx-js/tag": "^0.13.1-1",
-    "react": "^0.14.x || ^15.x || ^16.x"
+    "react": "^16.8.0"
   },
   "devDependencies": {
     "jest": "^23.3.0",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2"
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0"
   },
   "scripts": {
     "test": "jest"
