+++ packages/tag/package.json
@@ -31,12 +31,12 @@
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
     "jest": "^23.1.0",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
     "rimraf": "^2.6.3"
   },
   "peerDependencies": {
-    "react": "^0.14.x || ^15.x || ^16.x"
+    "react": "^16.8.0"
   },
   "scripts": {
     "build": "babel src -d dist",
