+++ packages/tag/src/index.js
@@ -1,2 +1,6 @@
 export {default as MDXTag} from './mdx-tag'
-export {default as MDXProvider, withMDXComponents} from './mdx-provider'
+export {
+  default as MDXContext,
+  MDXProvider,
+  withMDXComponents
+} from './mdx-context'
